feat: remove go ethereum package
