+++ node/.env.example
@@ -4,7 +4,6 @@ REDIS_HOST=
 REDIS_PORT=
 LISTEN_PORT=
 KLAYTN_PROVIDER_URL=
-ETH_PROVIDER_URL=
 SUBMISSION_PROXY_CONTRACT=
 DELEGATOR_URL=
 
@@ -13,7 +12,6 @@ CHAIN=
 
 # (optional) required if wallets table is empty
 KLAYTN_REPORTER_PK=
-ETH_REPORTER_PK=
 # (optional) required to run klaytn_helper test
 TEST_FEE_PAYER_PK=
