+++ node/.env.local
@@ -4,10 +4,10 @@ REDIS_HOST=localhost
 REDIS_PORT=6379
 LISTEN_PORT=10010
 KLAYTN_PROVIDER_URL=https://public-en-baobab.klaytn.net
-ETH_PROVIDER_URL=
+
 CHAIN=test
 KLAYTN_REPORTER_PK=
-ETH_REPORTER_PK=
+
 
 # this address is dummy contract in baobab
 SUBMISSION_PROXY_CONTRACT=0xc9fc64584acbf0f6beb1819b1b6cd3f5aefe34ea
