--- node/pkg/chain/eth/utils/types.go
@@ -1,5 +1,0 @@
-package utils
-
-const (
-	DEFAULT_GAS_LIMIT = uint64(1200000)
-)
